Move schema directory to top level module and fix tests
