+++ shared/src/commands/generate/components/instantiate.spec.ts
@@ -1,6 +1,6 @@
  
 
-import { SchemaDirectory } from '../schema-directory';
+import { SchemaDirectory } from '../../../schema-directory';
 import { instantiateGenericObject } from './instantiate';
 
 jest.mock('../../helper', () => {
@@ -14,7 +14,7 @@ jest.mock('../../helper', () => {
     };
 });
 
-jest.mock('../schema-directory');
+jest.mock('../../../schema-directory');
 
 let mockSchemaDir;
