+++ shared/src/commands/generate/components/metadata.spec.ts
@@ -1,4 +1,4 @@
-import { SchemaDirectory } from '../schema-directory';
+import { SchemaDirectory } from '../../../schema-directory';
 import { instantiateAllMetadata, instantiateMetadataObject } from './metadata';
 
 jest.mock('../../helper', () => {
@@ -12,7 +12,7 @@ jest.mock('../../helper', () => {
     };
 });
 
-jest.mock('../schema-directory');
+jest.mock('../../../schema-directory');
 
 let mockSchemaDir;
