+++ shared/src/commands/generate/components/relationship.ts
@@ -1,8 +1,8 @@
 /* eslint-disable  @typescript-eslint/no-explicit-any */
 
 import { initLogger } from '../../helper.js';
-import { SchemaDirectory } from '../schema-directory.js';
-import { appendPath } from '../util.js';
+import { SchemaDirectory } from '../../../schema-directory.js';
+import { appendPath } from '../../../util.js';
 import { instantiateGenericObject } from './instantiate.js';
