+++ shared/src/commands/generate/generate.spec.ts
@@ -14,7 +14,7 @@ jest.mock('../helper', () => {
     };
 });
 
-jest.mock('./schema-directory');
+jest.mock('../../schema-directory');
 
 jest.mock('../../consts', () => ({
     get CALM_META_SCHEMA_DIRECTORY() { return '../calm/draft/2024-10/meta'; }
